feat: добавлены тесты для пакета user

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -3,33 +3,28 @@ package ru.practicum.shareit.user.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.practicum.shareit.item.NotFoundException;
-import ru.practicum.shareit.user.EmailDuplicationException;
+import ru.practicum.shareit.NotFoundException;
 import ru.practicum.shareit.user.User;
 import ru.practicum.shareit.user.UserMapper;
-import ru.practicum.shareit.user.ValidationException;
-import ru.practicum.shareit.user.dao.UserDao;
 import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.repository.UserRepository;
 
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
-    private final UserDao userDao;
     private final UserRepository userRepository;
 
-    public User createUser(UserDto userDto) throws ValidationException, EmailDuplicationException {
-        return userDao.save(UserMapper.mapToUser(userDto));
-    }
-
+    @Transactional
     public User save(UserDto userDto) {
-        User user = userRepository.save(UserMapper.mapToUser(userDto));
-        return user;
+
+        return userRepository.save(UserMapper.mapToUser(userDto));
     }
 
+    @Transactional
     public User update(UserDto userDto, long userId) throws NotFoundException {
         if (userRepository.findById(userId).isPresent()) {
             User user = userRepository.findById(userId).get();
@@ -50,33 +45,33 @@ public class UserServiceImpl implements UserService {
         }
     }
 
-    public User updateUser(UserDto userDto, long userId)
-            throws NotFoundException, ValidationException, EmailDuplicationException {
-        try {
-            User user = new User();
-            if (userDto.getName() != null && userDto.getEmail() != null) {
-                user = UserMapper.mapToUser(userDto);
-                user.setId(userId);
-            } else {
-                user.setId(userId);
-                if (userDto.getName() != null) {
-                    user.setName(userDto.getName());
-                    user.setEmail(userDao.getUserById(userId).getEmail());
-                }
-                if (userDto.getEmail() != null) {
-                    user.setEmail(userDto.getEmail());
-                    user.setName(userDao.getUserById(userId).getName());
-                }
-            }
-            return userDao.update(user);
-        } catch (NotFoundException e) {
-            log.info(e.getMessage());
-            throw new NotFoundException(User.class.getName());
-        } catch (EmailDuplicationException e) {
-            log.info(e.getMessage());
-            throw new EmailDuplicationException(e.getMessage());
-        }
-    }
+//    public User updateUser(UserDto userDto, long userId)
+//            throws NotFoundException, ValidationException, EmailDuplicationException {
+//        try {
+//            User user = new User();
+//            if (userDto.getName() != null && userDto.getEmail() != null) {
+//                user = UserMapper.mapToUser(userDto);
+//                user.setId(userId);
+//            } else {
+//                user.setId(userId);
+//                if (userDto.getName() != null) {
+//                    user.setName(userDto.getName());
+//                    user.setEmail(userDao.getUserById(userId).getEmail());
+//                }
+//                if (userDto.getEmail() != null) {
+//                    user.setEmail(userDto.getEmail());
+//                    user.setName(userDao.getUserById(userId).getName());
+//                }
+//            }
+//            return userDao.update(user);
+//        } catch (NotFoundException e) {
+//            log.info(e.getMessage());
+//            throw new NotFoundException(User.class.getName());
+//        } catch (EmailDuplicationException e) {
+//            log.info(e.getMessage());
+//            throw new EmailDuplicationException(e.getMessage());
+//        }
+//    }
 
     public User getUserById(long userId) throws NotFoundException {
         try {
@@ -91,7 +86,11 @@ public class UserServiceImpl implements UserService {
     }
 
     public void deleteUser(long userId) {
-        userRepository.deleteById(userId);
+      if (userRepository.existsById(userId)) {
+          userRepository.deleteById(userId);
+      } else {
+          throw new NotFoundException(User.class.getName());
+      }
     }
 }
